+++ apps/web/app/api/billing/stripe-webhook/route.ts
@@ -1,15 +1,3 @@
-import { responses } from "@/app/lib/api/response";
-import { webhookHandler } from "@/modules/ee/billing/api/lib/stripeWebhook";
-import { headers } from "next/headers";
+import { POST } from "@/modules/ee/billing/api/route";
 
-export const POST = async (request: Request) => {
-  const body = await request.text();
-  const signature = headers().get("stripe-signature") as string;
-
-  const { status, message } = await webhookHandler(body, signature);
-
-  if (status != 200) {
-    return responses.badRequestResponse(message?.toString() || "Something went wrong");
-  }
-  return responses.successResponse({ message }, true);
-};
+export { POST };
