+++ apps/web/modules/ee/billing/actions.ts
@@ -3,15 +3,15 @@
 import { authenticatedActionClient } from "@/lib/utils/action-client";
 import { checkAuthorizationUpdated } from "@/lib/utils/action-client-middleware";
 import { getOrganizationIdFromEnvironmentId } from "@/lib/utils/helper";
+import { createCustomerPortalSession } from "@/modules/ee/billing/api/lib/create-customer-portal-session";
+import { createSubscription } from "@/modules/ee/billing/api/lib/create-subscription";
+import { isSubscriptionCancelled } from "@/modules/ee/billing/api/lib/is-subscription-cancelled";
 import { z } from "zod";
 import { STRIPE_PRICE_LOOKUP_KEYS } from "@formbricks/lib/constants";
 import { WEBAPP_URL } from "@formbricks/lib/constants";
 import { getOrganization } from "@formbricks/lib/organization/service";
 import { ZId } from "@formbricks/types/common";
 import { AuthorizationError, ResourceNotFoundError } from "@formbricks/types/errors";
-import { createCustomerPortalSession } from "./api/lib/createCustomerPortalSession";
-import { createSubscription } from "./api/lib/createSubscription";
-import { isSubscriptionCancelled } from "./api/lib/isSubscriptionCancelled";
 
 const ZUpgradePlanAction = z.object({
   environmentId: ZId,
